REFACTOR: Moved messages and tests of slashing module to own files.
