+++ pkg/indexer/decode/message_test.go
@@ -5,7 +5,6 @@ import (
 
 	"cosmossdk.io/math"
 	cosmosTypes "github.com/cosmos/cosmos-sdk/types"
-	cosmosSlashingTypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	cosmosStakingTypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/dipdup-io/celestia-indexer/internal/storage"
 	storageTypes "github.com/dipdup-io/celestia-indexer/internal/storage/types"
@@ -384,56 +383,6 @@ func TestDecodeMsg_SuccessOnMsgUndelegate(t *testing.T) {
 		Addresses: addressesExpected,
 	}
 
-	assert.NoError(t, err)
-	assert.Equal(t, uint64(0), dm.BlobsSize)
-	assert.Equal(t, msgExpected, dm.Msg)
-	assert.Equal(t, addressesExpected, dm.Addresses)
-}
-
-// MsgUnjail
-
-func createMsgUnjail() cosmosTypes.Msg {
-	m := cosmosSlashingTypes.MsgUnjail{
-		ValidatorAddr: "celestiavaloper170qq26qenw420ufd5py0r59kpg3tj2m7dqkpym",
-	}
-	return &m
-}
-
-func TestDecodeMsg_SuccessOnMsgUnjail(t *testing.T) {
-	m := createMsgUnjail()
-	blob, now := testsuite.EmptyBlock()
-	position := 0
-
-	dm, err := Message(m, blob.Height, blob.Block.Time, position, storageTypes.StatusSuccess)
-
-	addressesExpected := []storage.AddressWithType{
-		{
-			Type: storageTypes.MsgAddressTypeValidatorAddress,
-			Address: storage.Address{
-				Id:         0,
-				Height:     blob.Height,
-				LastHeight: blob.Height,
-				Address:    "celestiavaloper170qq26qenw420ufd5py0r59kpg3tj2m7dqkpym",
-				Hash:       []byte{0xf3, 0xc0, 0x5, 0x68, 0x19, 0x9b, 0xaa, 0xa7, 0xf1, 0x2d, 0xa0, 0x48, 0xf1, 0xd0, 0xb6, 0xa, 0x22, 0xb9, 0x2b, 0x7e},
-				Balance: storage.Balance{
-					Id:    0,
-					Total: decimal.Zero,
-				},
-			},
-		},
-	}
-
-	msgExpected := storage.Message{
-		Id:        0,
-		Height:    blob.Height,
-		Time:      now,
-		Position:  0,
-		Type:      storageTypes.MsgUnjail,
-		TxId:      0,
-		Data:      structs.Map(m),
-		Namespace: nil,
-		Addresses: addressesExpected,
-	}
 	assert.NoError(t, err)
 	assert.Equal(t, uint64(0), dm.BlobsSize)
 	assert.Equal(t, msgExpected, dm.Msg)
